Rebase and refresh to 2020-12-29; Scala 2.13 and JUnit changes
